Added overlap index and refactor of term coding
The overlap index has been calculated using the operations of  winners and no winners for each BPI question. This way it is possible to measure the similiarity between their answers since  correctly since this measure is not affected by the sizes of sets. Apart from this measure, the term coding file has been rename to key concept code which is the name used in the article